fix bug

+++ app.py
@@ -74,8 +74,8 @@ def webhook():
                 # user clicked/tapped "postback" button in earlier message
                 if messaging_event.get("postback"):
                     wantwatch = messaging_event["postback"]["payload"].split()
-                    px = int(wantwatch[0])
-                    py = int(wantwatch[1])
+                    px = float(wantwatch[0])
+                    py = float(wantwatch[1])
                     count = int(wantwatch[2])
                     search_scene(sender_id, px, py, count)
